Убрал пару ошибок

+++ EFMigrator/Database/User.cs
@@ -15,7 +15,7 @@ namespace Classes
         public string Password { get; set; }
         public uint Salary { get; set; }
         public string Email { get; set; }
-        public uint? FamilyId { get; set; }
+        public ushort? FamilyId { get; set; }
     }
